TheEmotionGame/requirements.txt
requests==2.24.0
telegram==0.0.1
msrest==0.6.19
pyodbc==4.0.30
azure_storage==0.37.0
Pillow==8.1.0
PyYAML==5.4.1
